make keysPerPage accessible via StoreBuilder, make TestRandomMapOps more rigorous

+++ h2/src/test/org/h2/test/store/TestRandomMapOps.java
@@ -7,12 +7,17 @@ package org.h2.test.store;
 
 import java.text.MessageFormat;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Random;
 import java.util.TreeMap;
+import org.h2.mvstore.Cursor;
 import org.h2.mvstore.MVMap;
 import org.h2.mvstore.MVStore;
 import org.h2.store.fs.FileUtils;
+import org.h2.test.TestAll;
 import org.h2.test.TestBase;
 
 /**
@@ -21,7 +26,9 @@ import org.h2.test.TestBase;
 public class TestRandomMapOps extends TestBase {
 
     private static final boolean LOG = false;
+    private final Random r = new Random();
     private int op;
+
 
     /**
      * Run just this test.
@@ -30,56 +37,65 @@ public class TestRandomMapOps extends TestBase {
      */
     public static void main(String... a) throws Exception {
         TestBase test = TestBase.createCaller().init();
-        test.config.big = true;
-        test.testFromMain();
+        TestAll config = test.config;
+        config.big = true;
+//        config.memory = true;
+
+        test.println(config.toString());
+        for (int i = 0; i < 10; i++) {
+            test.testFromMain();
+            test.println("Done pass #" + i);
+        }
     }
 
     @Override
     public void test() throws Exception {
-        testMap("memFS:randomOps.h3");
-        FileUtils.delete("memFS:randomOps.h3");
+        if (config.memory) {
+            testMap(null);
+        } else {
+            String fileName = "memFS:" + getTestName();
+            testMap(fileName);
+        }
     }
 
     private void testMap(String fileName) {
-        int best = Integer.MAX_VALUE;
-        int bestSeed = 0;
-        Throwable failException = null;
-        int size = getSize(100, 1000);
-        for (int seed = 0; seed < 100; seed++) {
-            FileUtils.delete(fileName);
-            Throwable ex = null;
+        int size = getSize(500, 3000);
+        long seed = 0;
+//        seed = System.currentTimeMillis();
+//        seed = -3407210256209708616L;
+        for (int cnt = 0; cnt < 100; cnt++) {
             try {
                 testOps(fileName, size, seed);
-                continue;
-            } catch (Exception | AssertionError e) {
-                ex = e;
+            } catch (Exception | AssertionError ex) {
+                println("seed:" + seed + " op:" + op + " " + ex);
+                throw ex;
+            } finally {
+                if (fileName != null) {
+                    FileUtils.delete(fileName);
+                }
             }
-            if (op < best) {
-                trace(seed);
-                bestSeed = seed;
-                best = op;
-                size = best;
-                failException = ex;
-                // System.out.println("seed:" + seed + " op:" + op + " " + ex);
-            }
-        }
-        if (failException != null) {
-            throw (AssertionError) new AssertionError("seed = " + bestSeed
-                    + " op = " + best).initCause(failException);
+            seed = r.nextLong();
         }
     }
 
-    private void testOps(String fileName, int size, int seed) {
-        FileUtils.delete(fileName);
-        MVStore s = openStore(fileName);
-        MVMap<Integer, byte[]> m = s.openMap("data");
-        Random r = new Random(seed);
+    private void testOps(String fileName, int loopCount, long seed) {
+        r.setSeed(seed);
         op = 0;
-        TreeMap<Integer, byte[]> map = new TreeMap<>();
-        for (; op < size; op++) {
-            int k = r.nextInt(100);
-            byte[] v = new byte[r.nextInt(10) * 10];
-            int type = r.nextInt(12);
+        MVStore s = openStore(fileName);
+        int keysPerPage = s.getKeysPerPage();
+        int keyRange = 2000;
+        MVMap<Integer, String> m = s.openMap("data");
+        TreeMap<Integer, String> map = new TreeMap<>();
+        int[] recentKeys = new int[2 * keysPerPage];
+        for (; op < loopCount; op++) {
+            int k = r.nextInt(3 * keyRange / 2);
+            if (k >= keyRange) {
+                k = recentKeys[k  % recentKeys.length];
+            } else {
+                recentKeys[op % recentKeys.length] = k;
+            }
+            String v = k + "_Value_" + op;
+            int type = r.nextInt(15);
             switch (type) {
             case 0:
             case 1:
@@ -100,23 +116,27 @@ public class TestRandomMapOps extends TestBase {
                 s.compact(90, 1024);
                 break;
             case 7:
-                log(op, k, v, "m.clear()");
-                m.clear();
-                map.clear();
+                if (op % 64 == 0) {
+                    log(op, k, v, "m.clear()");
+                    m.clear();
+                    map.clear();
+                }
                 break;
             case 8:
                 log(op, k, v, "s.commit()");
                 s.commit();
                 break;
             case 9:
-                log(op, k, v, "s.commit()");
-                s.commit();
-                log(op, k, v, "s.close()");
-                s.close();
-                log(op, k, v, "s = openStore(fileName)");
-                s = openStore(fileName);
-                log(op, k, v, "m = s.openMap(\"data\")");
-                m = s.openMap("data");
+                if (fileName != null) {
+                    log(op, k, v, "s.commit()");
+                    s.commit();
+                    log(op, k, v, "s.close()");
+                    s.close();
+                    log(op, k, v, "s = openStore(fileName)");
+                    s = openStore(fileName);
+                    log(op, k, v, "m = s.openMap(\"data\")");
+                    m = s.openMap("data");
+                }
                 break;
             case 10:
                 log(op, k, v, "s.commit()");
@@ -124,7 +144,30 @@ public class TestRandomMapOps extends TestBase {
                 log(op, k, v, "s.compactMoveChunks()");
                 s.compactMoveChunks();
                 break;
-            case 11:
+            case 11: {
+                int rangeSize = r.nextInt(2 * keysPerPage);
+                int step = r.nextBoolean() ? 1 : -1;
+                for (int i = 0; i < rangeSize; i++) {
+                    log(op, k, v, "m.put({0}, {1})");
+                    m.put(k, v);
+                    map.put(k, v);
+                    k += step;
+                    v = k + "_Value_" + op;
+                }
+                break;
+            }
+            case 12: {
+                int rangeSize = r.nextInt(2 * keysPerPage);
+                int step = r.nextBoolean() ? 1 : -1;
+                for (int i = 0; i < rangeSize; i++) {
+                    log(op, k, v, "m.remove({0})");
+                    m.remove(k);
+                    map.remove(k);
+                    k += step;
+                }
+                break;
+            }
+            default:
                 log(op, k, v, "m.getKeyIndex({0})");
                 ArrayList<Integer> keyList = new ArrayList<>(map.keySet());
                 int index = Collections.binarySearch(keyList, k, null);
@@ -136,7 +179,7 @@ public class TestRandomMapOps extends TestBase {
                 }
                 break;
             }
-            assertEqualsMapValues(map.get(k), m.get(k));
+            assertEquals(map.get(k), m.get(k));
             assertEquals(map.ceilingKey(k), m.ceilingKey(k));
             assertEquals(map.floorKey(k), m.floorKey(k));
             assertEquals(map.higherKey(k), m.higherKey(k));
@@ -147,25 +190,81 @@ public class TestRandomMapOps extends TestBase {
                 assertEquals(map.firstKey(), m.firstKey());
                 assertEquals(map.lastKey(), m.lastKey());
             }
+
+            int from = r.nextBoolean() ? r.nextInt(keyRange) : k + r.nextInt(2 * keysPerPage) - keysPerPage;
+            int to = r.nextBoolean() ? r.nextInt(keyRange) : from + r.nextInt(2 * keysPerPage) - keysPerPage;
+
+            Cursor<Integer, String> cursor;
+            Collection<Map.Entry<Integer, String>> entrySet;
+            String msg;
+            if (from <= to) {
+                msg = "(" + from + ", null)";
+                cursor = m.cursor(from, null, false);
+                entrySet = map.tailMap(from).entrySet();
+                assertEquals(msg, entrySet, cursor);
+
+                msg = "(null, " + from + ")";
+                cursor = m.cursor(null, from, false);
+                entrySet = map.headMap(from + 1).entrySet();
+                assertEquals(msg, entrySet, cursor);
+
+                msg = "(" + from + ", " + to + ")";
+                cursor = m.cursor(from, to, false);
+                entrySet = map.subMap(from, to + 1).entrySet();
+                assertEquals(msg, entrySet, cursor);
+            }
+
+            if (from >= to) {
+                msg = "rev (" + from + ", null)";
+                cursor = m.cursor(from, null, true);
+                entrySet = reverse(map.headMap(from + 1).entrySet());
+                assertEquals(msg, entrySet, cursor);
+
+                msg = "rev (null, "+from+")";
+                cursor = m.cursor(null, from, true);
+                entrySet = reverse(map.tailMap(from).entrySet());
+                assertEquals(msg, entrySet, cursor);
+
+                msg = "rev (" + from + ", " + to + ")";
+                cursor = m.cursor(from, to, true);
+                entrySet = reverse(map.subMap(to, from + 1).entrySet());
+                assertEquals(msg, entrySet, cursor);
+            }
         }
         s.close();
     }
 
-    private static MVStore openStore(String fileName) {
-        MVStore s = new MVStore.Builder().fileName(fileName).
-                pageSplitSize(50).autoCommitDisabled().open();
-        s.setRetentionTime(1000);
-        return s;
+    private <K,V> Collection<Map.Entry<K,V>> reverse(Collection<Map.Entry<K,V>> entrySet) {
+        ArrayList<Map.Entry<K,V>> list = new ArrayList<>(entrySet);
+        Collections.reverse(list);
+        entrySet = list;
+        return entrySet;
     }
 
-    private void assertEqualsMapValues(byte[] x, byte[] y) {
-        if (x == null || y == null) {
-            if (x != y) {
-                assertTrue(x == y);
-            }
-        } else {
-            assertEquals(x.length, y.length);
+    private <K,V> void assertEquals(String msg, Iterable<Map.Entry<K, V>> entrySet, Cursor<K, V> cursor) {
+        int cnt = 0;
+        for (Map.Entry<K,V> entry : entrySet) {
+            String message = msg + " " + cnt;
+            assertTrue(message, cursor.hasNext());
+            assertEquals(message, entry.getKey(), cursor.next());
+            assertEquals(message, entry.getKey(), cursor.getKey());
+            assertEquals(message, entry.getValue(), cursor.getValue());
+            ++cnt;
         }
+        assertFalse(msg, cursor.hasNext());
+    }
+
+    public void assertEquals(String message, Object expected, Object actual) {
+        if (!Objects.equals(expected, actual)) {
+            fail(message + " expected: " + expected + " actual: " + actual);
+        }
+    }
+
+    private static MVStore openStore(String fileName) {
+        MVStore s = new MVStore.Builder().fileName(fileName)
+                .keysPerPage(7).autoCommitDisabled().open();
+        s.setRetentionTime(1000);
+        return s;
     }
 
     /**
@@ -176,10 +275,9 @@ public class TestRandomMapOps extends TestBase {
      * @param v the value
      * @param msg the message
      */
-    private static void log(int op, int k, byte[] v, String msg) {
+    private static void log(int op, int k, String v, String msg) {
         if (LOG) {
-            msg = MessageFormat.format(msg, k,
-                    v == null ? null : "new byte[" + v.length + "]");
+            msg = MessageFormat.format(msg, k, v);
             System.out.println(msg + "; // op " + op);
         }
     }
